+++ pkgs/firehose/lib/src/health/coverage.dart
@@ -20,6 +20,7 @@ class Coverage {
   final List<Glob> ignoredPackages;
   final Directory directory;
   final List<String> experiments;
+  final String dartExecutable;
 
   Coverage(
     this.coverageWeb,
@@ -27,17 +28,12 @@ class Coverage {
     this.ignoredPackages,
     this.directory,
     this.experiments,
+    this.dartExecutable,
   );
-
-  Future<CoverageResult> compareCoverages(
-      GithubApi github, Directory base) async {
-    var files = await github.listFilesForPR(directory, ignoredFiles);
-    return compareCoveragesFor(files, base);
-  }
 
   CoverageResult compareCoveragesFor(List<GitFile> files, Directory base) {
     var repository = Repository(directory);
-    var packages = repository.locatePackages(ignoredPackages);
+    var packages = repository.locatePackages(ignore: ignoredPackages);
     print('Found packages $packages at $directory');
 
     var filesOfInterest = files
@@ -49,7 +45,7 @@ class Coverage {
     print('The files of interest are $filesOfInterest');
 
     var baseRepository = Repository(base);
-    var basePackages = baseRepository.locatePackages(ignoredFiles);
+    var basePackages = baseRepository.locatePackages(ignore: ignoredFiles);
     print('Found packages $basePackages at $base');
 
     var changedPackages = packages
@@ -105,7 +101,7 @@ class Coverage {
         print('''
 Get coverage for ${package.name} by running coverage in ${package.directory.path}''');
         Process.runSync(
-          'dart',
+          dartExecutable,
           [
             if (experiments.isNotEmpty)
               '--enable-experiment=${experiments.join(',')}',
@@ -117,7 +113,7 @@ Get coverage for ${package.name} by running coverage in ${package.directory.path
         if (coverageWeb) {
           print('Run tests with coverage for web');
           var resultChrome = Process.runSync(
-            'dart',
+            dartExecutable,
             [
               if (experiments.isNotEmpty)
                 '--enable-experiment=${experiments.join(',')}',
@@ -136,7 +132,7 @@ Get coverage for ${package.name} by running coverage in ${package.directory.path
 
         print('Run tests with coverage for vm');
         var resultVm = Process.runSync(
-          'dart',
+          dartExecutable,
           [
             if (experiments.isNotEmpty)
               '--enable-experiment=${experiments.join(',')}',
@@ -152,7 +148,7 @@ Get coverage for ${package.name} by running coverage in ${package.directory.path
 
         print('Compute coverage from runs');
         var resultLcov = Process.runSync(
-          'dart',
+          dartExecutable,
           [
             'pub',
             'global',