+++ pkgs/firehose/test/coverage_test.dart
@@ -43,7 +43,7 @@ void main() {
 }
 
 class FakeHealth extends Coverage {
-  FakeHealth() : super(true, [], [], Directory.current, []);
+  FakeHealth() : super(true, [], [], Directory.current, [], 'dart');
 
   @override
   Map<String, double> getCoverage(Package? package) {